refactor: split operator management pages
