+++ src/cook-web/src/app/admin/operations/promotions/page.tsx
@@ -1,8 +1,8 @@
 'use client';
 
 import React, { useCallback, useEffect, useRef, useState } from 'react';
-import { Trans, useTranslation } from 'react-i18next';
-import { CalendarIcon, Plus, X } from 'lucide-react';
+import { useTranslation } from 'react-i18next';
+import { Plus } from 'lucide-react';
 import api from '@/api';
 import AdminClearableInput from '@/app/admin/components/AdminClearableInput';
 import AdminDateRangeFilter from '@/app/admin/components/AdminDateRangeFilter';
@@ -10,49 +10,19 @@ import AdminFilter from '@/app/admin/components/AdminFilter';
 import AdminBreadcrumb from '@/app/admin/components/AdminBreadcrumb';
 import AdminTitle from '@/app/admin/components/AdminTitle';
 import AdminTableShell from '@/app/admin/components/AdminTableShell';
-import AdminTooltipText from '@/app/admin/components/AdminTooltipText';
 import AdminRowActions from '@/app/admin/components/AdminRowActions';
-import AdminTimeSelect from '@/app/admin/components/AdminTimeSelect';
-import {
-  ADMIN_TABLE_HEADER_CELL_CENTER_CLASS,
-  ADMIN_TABLE_RESIZE_HANDLE_CLASS,
-  getAdminStickyRightCellClass,
-  getAdminStickyRightHeaderClass,
-} from '@/app/admin/components/adminTableStyles';
+import { ADMIN_TABLE_RESIZE_HANDLE_CLASS } from '@/app/admin/components/adminTableStyles';
 import { useAdminResizableColumns } from '@/app/admin/hooks/useAdminResizableColumns';
 import { formatAdminUtcDateTime } from '@/app/admin/lib/dateTime';
 import type {
   AdminPromotionCampaignItem,
-  AdminPromotionCampaignRedemptionItem,
   AdminPromotionCouponCodeItem,
   AdminPromotionCouponItem,
-  AdminPromotionCouponUsageItem,
   AdminPromotionListResponse,
 } from '@/app/admin/operations/operation-promotion-types';
 import useOperatorGuard from '@/app/admin/operations/useOperatorGuard';
 import ErrorDisplay from '@/components/ErrorDisplay';
-import {
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-} from '@/components/ui/AlertDialog';
-import { Badge } from '@/components/ui/Badge';
 import { Button } from '@/components/ui/Button';
-import {
-  Dialog,
-  DialogContent,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-} from '@/components/ui/Dialog';
-import { Input } from '@/components/ui/Input';
-import { Label } from '@/components/ui/Label';
-import { Calendar } from '@/components/ui/Calendar';
 import {
   Select,
   SelectContent,
@@ -69,1987 +39,59 @@ import {
   TableRow,
 } from '@/components/ui/Table';
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/Tabs';
-import { Textarea } from '@/components/ui/Textarea';
 import { showDefaultToast, showErrorToast } from '@/hooks/useToast';
 import { cn } from '@/lib/utils';
-
-type PromotionTab = 'coupons' | 'campaigns';
-
-type CouponFilters = {
-  keyword: string;
-  name: string;
-  course_query: string;
-  usage_type: string;
-  ops_state: string;
-  discount_type: string;
-  status: string;
-  start_time: string;
-  end_time: string;
-};
-
-type CampaignFilters = {
-  keyword: string;
-  course_query: string;
-  apply_type: string;
-  channel: string;
-  discount_type: string;
-  status: string;
-  start_time: string;
-  end_time: string;
-};
-
-type CouponFormState = {
-  name: string;
-  code: string;
-  usage_type: string;
-  discount_type: string;
-  value: string;
-  total_count: string;
-  scope_type: string;
-  shifu_bid: string;
-  start_at: string;
-  end_at: string;
-  enabled: string;
-};
-
-type CampaignFormState = {
-  name: string;
-  apply_type: string;
-  shifu_bid: string;
-  discount_type: string;
-  value: string;
-  start_at: string;
-  end_at: string;
-  description: string;
-  channel: string;
-  enabled: string;
-};
-
-type ErrorState = { message: string } | null;
-type PromotionStatusChangeTarget =
-  | {
-      entityType: 'coupon';
-      enabling: boolean;
-      item: AdminPromotionCouponItem;
-    }
-  | {
-      entityType: 'campaign';
-      enabling: boolean;
-      item: AdminPromotionCampaignItem;
-    };
-
-const PAGE_SIZE = 20;
-const EMPTY_VALUE = '--';
-const ALL_OPTION_VALUE = '__all__';
-const PROMOTION_EXPIRING_SOON_DAYS = 7;
-const COLUMN_MIN_WIDTH = 90;
-const COLUMN_MAX_WIDTH = 420;
-const COUPON_COLUMN_WIDTH_STORAGE_KEY = 'adminPromotionCouponsColumnWidths';
-const CAMPAIGN_COLUMN_WIDTH_STORAGE_KEY = 'adminPromotionCampaignsColumnWidths';
-const COUPON_DEFAULT_COLUMN_WIDTHS = {
-  name: 200,
-  status: 110,
-  usageType: 120,
-  discountRule: 120,
-  code: 180,
-  scope: 120,
-  course: 240,
-  activeTime: 260,
-  usageProgress: 110,
-  codesEntry: 110,
-  couponBid: 220,
-  updatedAt: 170,
-  createdAt: 170,
-  action: 120,
-} as const;
-const CAMPAIGN_DEFAULT_COLUMN_WIDTHS = {
-  name: 200,
-  status: 110,
-  applyType: 120,
-  channel: 180,
-  course: 240,
-  discountRule: 120,
-  campaignTime: 280,
-  appliedOrderCount: 130,
-  promoBid: 220,
-  updatedAt: 170,
-  createdAt: 170,
-  action: 120,
-} as const;
-const PROMOTION_CODE_DIALOG_COLUMN_COUNT = 4;
-const PROMOTION_REDEMPTION_DIALOG_COLUMN_COUNT = 4;
-const PROMOTION_USAGE_DIALOG_COLUMN_COUNT = {
-  default: 4,
-  withCourse: 5,
-} as const;
-const SINGLE_SELECT_ITEM_CLASS =
-  'pl-3 data-[state=checked]:bg-muted data-[state=checked]:text-foreground [&>span:first-child]:hidden';
-const TABLE_HEAD_CLASS = ADMIN_TABLE_HEADER_CELL_CENTER_CLASS;
-const TABLE_ACTION_HEAD_CLASS = getAdminStickyRightHeaderClass('text-center');
-const TABLE_CELL_CLASS =
-  'border-r border-border last:border-r-0 whitespace-nowrap overflow-hidden text-ellipsis text-center';
-const TABLE_LAST_CELL_CLASS =
-  'whitespace-nowrap overflow-hidden text-ellipsis text-center';
-const TABLE_ACTION_CELL_CLASS = getAdminStickyRightCellClass(
-  'whitespace-nowrap text-center',
-);
-type CouponColumnKey = keyof typeof COUPON_DEFAULT_COLUMN_WIDTHS;
-type CampaignColumnKey = keyof typeof CAMPAIGN_DEFAULT_COLUMN_WIDTHS;
-
-const createDefaultCouponFilters = (): CouponFilters => ({
-  keyword: '',
-  name: '',
-  course_query: '',
-  usage_type: '',
-  ops_state: '',
-  discount_type: '',
-  status: '',
-  start_time: '',
-  end_time: '',
-});
-
-const createDefaultCampaignFilters = (): CampaignFilters => ({
-  keyword: '',
-  course_query: '',
-  apply_type: '',
-  channel: '',
-  discount_type: '',
-  status: '',
-  start_time: '',
-  end_time: '',
-});
-
-const createDefaultCouponForm = (): CouponFormState => ({
-  name: '',
-  code: '',
-  usage_type: '',
-  discount_type: '',
-  value: '',
-  total_count: '',
-  scope_type: 'single_course',
-  shifu_bid: '',
-  start_at: '',
-  end_at: '',
-  enabled: 'true',
-});
-
-function normalizePromotionFormDateTimeValue(value?: string) {
-  const formatted = formatAdminUtcDateTime(value || '');
-  return formatted || value || '';
-}
-
-const createCouponFormFromItem = (
-  item: AdminPromotionCouponItem,
-): CouponFormState => ({
-  name: item.name || '',
-  code: item.code || '',
-  usage_type: String(item.usage_type || ''),
-  discount_type: String(item.discount_type || ''),
-  value: item.value || '',
-  total_count: String(item.total_count || ''),
-  scope_type: item.scope_type || 'single_course',
-  shifu_bid: item.shifu_bid || '',
-  start_at: normalizePromotionFormDateTimeValue(item.start_at),
-  end_at: normalizePromotionFormDateTimeValue(item.end_at),
-  enabled: 'true',
-});
-
-const createDefaultCampaignForm = (): CampaignFormState => ({
-  name: '',
-  apply_type: '',
-  shifu_bid: '',
-  discount_type: '',
-  value: '',
-  start_at: '',
-  end_at: '',
-  description: '',
-  channel: '',
-  enabled: 'true',
-});
-
-const createCampaignFormFromItem = (
-  item: AdminPromotionCampaignItem,
-  description: string,
-): CampaignFormState => ({
-  name: item.name || '',
-  apply_type: String(item.apply_type || ''),
-  shifu_bid: item.shifu_bid || '',
-  discount_type: String(item.discount_type || ''),
-  value: item.value || '',
-  start_at: normalizePromotionFormDateTimeValue(item.start_at),
-  end_at: normalizePromotionFormDateTimeValue(item.end_at),
-  description: description || '',
-  channel: item.channel || '',
-  enabled: 'true',
-});
-
-const SectionCard = ({
-  title,
-  action,
-  children,
-}: React.PropsWithChildren<{ title: string; action?: React.ReactNode }>) => (
-  <div className='rounded-xl border border-border bg-white p-5 shadow-sm'>
-    {title || action ? (
-      <div
-        className={cn(
-          'mb-4 flex items-center gap-4',
-          title ? 'justify-between' : 'justify-start',
-        )}
-      >
-        {title ? (
-          <h2 className='text-base font-semibold text-foreground'>{title}</h2>
-        ) : null}
-        {action}
-      </div>
-    ) : null}
-    {children}
-  </div>
-);
-
-const renderTimeRange = (startAt?: string, endAt?: string) => {
-  const start = formatAdminUtcDateTime(startAt || '');
-  const end = formatAdminUtcDateTime(endAt || '');
-  if (!start && !end) return EMPTY_VALUE;
-  return `${start || EMPTY_VALUE} ~ ${end || EMPTY_VALUE}`;
-};
-
-const downloadExcelCompatibleCodesFile = (
-  fileName: string,
-  headerLabel: string,
-  codes: string[],
-) => {
-  const tableRows = codes
-    .map(
-      code =>
-        `<tr><td style="mso-number-format:'\\@';">${String(code)
-          .replace(/&/g, '&amp;')
-          .replace(/</g, '&lt;')
-          .replace(/>/g, '&gt;')}</td></tr>`,
-    )
-    .join('');
-  const html = `<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="UTF-8" />
-  </head>
-  <body>
-    <table>
-      <thead>
-        <tr><th>${headerLabel
-          .replace(/&/g, '&amp;')
-          .replace(/</g, '&lt;')
-          .replace(/>/g, '&gt;')}</th></tr>
-      </thead>
-      <tbody>${tableRows}</tbody>
-    </table>
-  </body>
-</html>`;
-  const blob = new Blob(['\ufeff', html], {
-    type: 'application/vnd.ms-excel;charset=utf-8;',
-  });
-  const url = window.URL.createObjectURL(blob);
-  const anchor = document.createElement('a');
-  anchor.href = url;
-  anchor.download = fileName;
-  document.body.appendChild(anchor);
-  anchor.click();
-  document.body.removeChild(anchor);
-  window.URL.revokeObjectURL(url);
-};
-
-const renderRuleLabel = (discountTypeKey: string, value: string) => {
-  if (discountTypeKey.endsWith('percent')) {
-    return `${value}%`;
-  }
-  return `- ¥${value}`;
-};
-
-const toPromotionRelativeKey = (key?: string) => {
-  if (!key) {
-    return '';
-  }
-  return key.startsWith('module.operationsPromotion.')
-    ? key.replace('module.operationsPromotion.', '')
-    : key;
-};
-
-const resolveCouponUsageTypeLabel = (
-  tPromotion: (key: string) => string,
-  usageType: number | string,
-  usageTypeKey?: string,
-) => {
-  if (usageTypeKey) {
-    const translated = tPromotion(toPromotionRelativeKey(usageTypeKey));
-    if (translated && translated !== usageTypeKey) {
-      return translated;
-    }
-  }
-  if (Number(usageType) === 801) {
-    return tPromotion('usageType.generic');
-  }
-  if (Number(usageType) === 802) {
-    return tPromotion('usageType.singleUse');
-  }
-  return EMPTY_VALUE;
-};
-
-const resolveCouponScopeLabel = (
-  tPromotion: (key: string) => string,
-  scopeType?: string,
-) => {
-  if (scopeType === 'all_courses') {
-    return tPromotion('scope.allCourses');
-  }
-  if (scopeType === 'single_course') {
-    return tPromotion('scope.singleCourse');
-  }
-  return EMPTY_VALUE;
-};
-
-const PROMOTION_STATUS_FALLBACK_KEYS: Record<string, string> = {
-  active: 'status.active',
-  ended: 'status.ended',
-  expired: 'status.expired',
-  inactive: 'status.inactive',
-  not_started: 'status.notStarted',
-  upcoming: 'status.notStarted',
-};
-
-const resolvePromotionStatusLabel = (
-  tPromotion: (key: string) => string,
-  statusKey?: string,
-  status?: string,
-) => {
-  const fallbackKey = status ? PROMOTION_STATUS_FALLBACK_KEYS[status] : '';
-  const translationKey = statusKey
-    ? toPromotionRelativeKey(statusKey)
-    : fallbackKey;
-  if (!translationKey) {
-    return EMPTY_VALUE;
-  }
-  const translated = tPromotion(translationKey);
-  return translated && translated !== translationKey ? translated : EMPTY_VALUE;
-};
-
-const resolvePromotionStatusBadgeClassName = (status?: string) => {
-  switch (status) {
-    case 'active':
-      return 'border-emerald-200 bg-emerald-50 text-emerald-700 hover:bg-emerald-50';
-    case 'not_started':
-      return 'border-sky-200 bg-sky-50 text-sky-700 hover:bg-sky-50';
-    case 'inactive':
-      return 'border-slate-200 bg-slate-100 text-slate-700 hover:bg-slate-100';
-    case 'expired':
-    case 'ended':
-      return 'border-orange-200 bg-orange-50 text-orange-700 hover:bg-orange-50';
-    default:
-      return 'border-slate-200 bg-slate-50 text-slate-700 hover:bg-slate-50';
-  }
-};
-
-const renderPromotionStatusBadge = ({
-  tPromotion,
-  statusKey,
-  status,
-}: {
-  tPromotion: (key: string) => string;
-  statusKey?: string;
-  status?: string;
-}) => (
-  <Badge
-    variant='outline'
-    className={cn(
-      'rounded-full px-2 py-0.5 text-xs font-medium',
-      resolvePromotionStatusBadgeClassName(status),
-    )}
-  >
-    {resolvePromotionStatusLabel(tPromotion, statusKey, status)}
-  </Badge>
-);
-
-const resolveCampaignApplyTypeLabel = (
-  tPromotion: (key: string) => string,
-  applyType: number | string,
-) => {
-  if (Number(applyType) === 2101) {
-    return tPromotion('campaign.applyTypeAuto');
-  }
-  if (Number(applyType) === 2102) {
-    return tPromotion('campaign.applyTypeEvent');
-  }
-  if (Number(applyType) === 2103) {
-    return tPromotion('campaign.applyTypeManual');
-  }
-  return EMPTY_VALUE;
-};
-
-const canEditCampaignStrategyFields = (item: AdminPromotionCampaignItem) => {
-  const startAt = parseLocalDateTimeInput(item.start_at || '');
-  if (!startAt) {
-    return false;
-  }
-  return startAt.getTime() > Date.now() && !item.has_redemptions;
-};
-
-const canEnableCouponItem = (item: AdminPromotionCouponItem) => {
-  const endAt = parseDateValue(item.end_at || '');
-  if (endAt && endAt.getTime() < Date.now()) {
-    return false;
-  }
-  return Number(item.used_count || 0) < Number(item.total_count || 0);
-};
-
-const canEnableCampaignItem = (item: AdminPromotionCampaignItem) => {
-  const endAt = parseDateValue(item.end_at || '');
-  return !endAt || endAt.getTime() >= Date.now();
-};
-
-const shouldShowCouponStatusToggle = (item: AdminPromotionCouponItem) =>
-  item.computed_status !== 'inactive' || canEnableCouponItem(item);
-
-const shouldShowCampaignStatusToggle = (item: AdminPromotionCampaignItem) =>
-  item.computed_status !== 'inactive' || canEnableCampaignItem(item);
-
-const renderUserLabel = (
-  item:
-    | AdminPromotionCouponUsageItem
-    | AdminPromotionCouponCodeItem
-    | AdminPromotionCampaignRedemptionItem,
-) => {
-  return item.user_mobile || item.user_email || item.user_bid || EMPTY_VALUE;
-};
-
-const parseLocalDateTimeInput = (value: string) => {
-  const normalized = String(value || '').trim();
-  if (!normalized) {
-    return null;
-  }
-  const parsed = new Date(
-    normalized.includes(' ') ? normalized.replace(' ', 'T') : normalized,
-  );
-  if (Number.isNaN(parsed.getTime())) {
-    return null;
-  }
-  return parsed;
-};
-
-const formatDateValue = (date: Date) => {
-  const year = date.getFullYear();
-  const month = String(date.getMonth() + 1).padStart(2, '0');
-  const day = String(date.getDate()).padStart(2, '0');
-  return `${year}-${month}-${day}`;
-};
-
-const formatTimeValue = (date: Date) => {
-  const hours = String(date.getHours()).padStart(2, '0');
-  const minutes = String(date.getMinutes()).padStart(2, '0');
-  return `${hours}:${minutes}`;
-};
-
-const parseDateValue = (value: string) => {
-  if (!value) {
-    return undefined;
-  }
-  const parsed = new Date(
-    String(value).includes(' ')
-      ? String(value).replace(' ', 'T')
-      : String(value),
-  );
-  if (Number.isNaN(parsed.getTime())) {
-    return undefined;
-  }
-  return parsed;
-};
-
-const isPromotionExpiringSoon = (endAt?: string) => {
-  const endDate = parseDateValue(endAt || '');
-  if (!endDate) {
-    return false;
-  }
-  const now = new Date();
-  const diff = endDate.getTime() - now.getTime();
-  return (
-    diff >= 0 && diff <= PROMOTION_EXPIRING_SOON_DAYS * 24 * 60 * 60 * 1000
-  );
-};
-
-const isCouponUsedUp = (item: AdminPromotionCouponItem) =>
-  Number(item.used_count || 0) >= Number(item.total_count || 0);
-
-const renderCouponAttentionBadges = (
-  item: AdminPromotionCouponItem,
-  tPromotion: (key: string) => string,
-) => {
-  if (item.computed_status !== 'active') {
-    return [];
-  }
-
-  if (isCouponUsedUp(item)) {
-    return [
-      <Badge
-        key='used-up'
-        variant='outline'
-        className='rounded-full border-rose-200 bg-rose-50 px-2 py-0.5 text-xs font-medium text-rose-700 hover:bg-rose-50'
-      >
-        {tPromotion('opsState.usedUp')}
-      </Badge>,
-    ];
-  }
-
-  if (isPromotionExpiringSoon(item.end_at)) {
-    return [
-      <Badge
-        key='expiring-soon'
-        variant='outline'
-        className='rounded-full border-amber-200 bg-amber-50 px-2 py-0.5 text-xs font-medium text-amber-800 hover:bg-amber-50'
-      >
-        {tPromotion('opsState.expiringSoon')}
-      </Badge>,
-    ];
-  }
-
-  return [];
-};
-
-const DEFAULT_START_TIME = '00:00';
-const DEFAULT_END_TIME = '23:59';
-
-const resolveDateTimeParts = (
-  value: string,
-  defaultTime: string,
-): { date: string; time: string } => {
-  const parsed = parseDateValue(value);
-  if (!parsed) {
-    return { date: '', time: defaultTime };
-  }
-  return {
-    date: formatDateValue(parsed),
-    time: formatTimeValue(parsed),
-  };
-};
-
-const combineDateAndTime = (dateValue: string, timeValue: string) => {
-  const normalizedDate = String(dateValue || '').trim();
-  if (!normalizedDate) {
-    return '';
-  }
-  const normalizedTime = String(timeValue || '').trim() || DEFAULT_START_TIME;
-  return `${normalizedDate} ${normalizedTime}:00`;
-};
-
-const isPositiveIntegerString = (value: string) => /^\d+$/.test(value.trim());
-
-const renderTooltipText = (text?: string, className?: string) => (
-  <AdminTooltipText
-    text={text}
-    emptyValue={EMPTY_VALUE}
-    className={className}
-  />
-);
-
-const FormField = ({
-  label,
-  children,
-}: React.PropsWithChildren<{ label: string }>) => (
-  <div className='space-y-2'>
-    <Label className='text-sm font-medium text-foreground'>{label}</Label>
-    {children}
-  </div>
-);
-
-const PromotionStatusConfirmDialog = ({
-  changeTarget,
-  submitting,
-  onOpenChange,
-  onConfirm,
-}: {
-  changeTarget: PromotionStatusChangeTarget | null;
-  submitting: boolean;
-  onOpenChange: (open: boolean) => void;
-  onConfirm: () => Promise<void>;
-}) => {
-  const { t } = useTranslation();
-  const { t: tPromotion } = useTranslation('module.operationsPromotion');
-  const itemName = React.useMemo(() => {
-    if (!changeTarget) {
-      return '';
-    }
-    return (
-      changeTarget.item.name ||
-      (changeTarget.entityType === 'coupon'
-        ? changeTarget.item.coupon_bid
-        : changeTarget.item.promo_bid)
-    );
-  }, [changeTarget]);
-
-  return (
-    <AlertDialog
-      open={Boolean(changeTarget)}
-      onOpenChange={onOpenChange}
-    >
-      <AlertDialogContent className='sm:max-w-[440px]'>
-        <AlertDialogHeader>
-          <AlertDialogTitle>
-            {changeTarget?.enabling
-              ? tPromotion('messages.enableConfirmTitle')
-              : tPromotion('messages.disableConfirmTitle')}
-          </AlertDialogTitle>
-          <AlertDialogDescription className='text-left text-sm text-muted-foreground'>
-            {changeTarget ? (
-              <Trans
-                ns='module.operationsPromotion'
-                i18nKey={
-                  changeTarget.enabling
-                    ? 'messages.enableConfirmDescription'
-                    : 'messages.disableConfirmDescription'
-                }
-                values={{ name: itemName }}
-                components={{
-                  strong: <span className='font-semibold text-foreground' />,
-                }}
-              />
-            ) : null}
-          </AlertDialogDescription>
-        </AlertDialogHeader>
-        <AlertDialogFooter>
-          <AlertDialogCancel disabled={submitting}>
-            {t('common.core.cancel')}
-          </AlertDialogCancel>
-          <AlertDialogAction
-            disabled={submitting}
-            onClick={event => {
-              event.preventDefault();
-              void onConfirm();
-            }}
-          >
-            {changeTarget?.enabling
-              ? tPromotion('actions.confirmEnable')
-              : tPromotion('actions.confirmDisable')}
-          </AlertDialogAction>
-        </AlertDialogFooter>
-      </AlertDialogContent>
-    </AlertDialog>
-  );
-};
-
-const PromotionDateTimePicker = ({
-  value,
-  placeholder,
-  resetLabel,
-  clearLabel,
-  timeLabel,
-  defaultTime,
-  minDateTime,
-  maxDateTime,
-  onChange,
-}: {
-  value: string;
-  placeholder: string;
-  resetLabel: string;
-  clearLabel: string;
-  timeLabel: string;
-  defaultTime: string;
-  minDateTime?: string;
-  maxDateTime?: string;
-  onChange: (value: string) => void;
-}) => {
-  const { t } = useTranslation();
-  const [open, setOpen] = useState(false);
-  const [draftDate, setDraftDate] = useState<Date | undefined>(undefined);
-  const [draftTime, setDraftTime] = useState(defaultTime);
-  const selectedDate = React.useMemo(() => parseDateValue(value), [value]);
-  const minDate = React.useMemo(
-    () => parseDateValue(minDateTime || ''),
-    [minDateTime],
-  );
-  const maxDate = React.useMemo(
-    () => parseDateValue(maxDateTime || ''),
-    [maxDateTime],
-  );
-  const timeParts = React.useMemo(
-    () => resolveDateTimeParts(value, defaultTime),
-    [defaultTime, value],
-  );
-  const minParts = React.useMemo(
-    () => resolveDateTimeParts(minDateTime || '', DEFAULT_START_TIME),
-    [minDateTime],
-  );
-  const maxParts = React.useMemo(
-    () => resolveDateTimeParts(maxDateTime || '', DEFAULT_END_TIME),
-    [maxDateTime],
-  );
-  const minDateKey = minDate ? formatDateValue(minDate) : '';
-  const maxDateKey = maxDate ? formatDateValue(maxDate) : '';
-  const resolveInitialCalendarMonth = React.useCallback(
-    () => selectedDate || minDate || maxDate || new Date(),
-    [maxDate, minDate, selectedDate],
-  );
-  const [calendarMonth, setCalendarMonth] = useState<Date>(
-    resolveInitialCalendarMonth,
-  );
-  const hasValue = Boolean(value);
-  const label = selectedDate
-    ? `${formatDateValue(selectedDate)} ${timeParts.time}`
-    : placeholder;
-  const draftDateKey = draftDate ? formatDateValue(draftDate) : '';
-  const minTime =
-    draftDateKey && draftDateKey === minDateKey ? minParts.time : undefined;
-  const maxTime =
-    draftDateKey && draftDateKey === maxDateKey ? maxParts.time : undefined;
-  const isDraftTimeOutOfRange =
-    (Boolean(minTime) && draftTime < String(minTime)) ||
-    (Boolean(maxTime) && draftTime > String(maxTime));
-  const isDayDisabled = React.useCallback(
-    (date: Date) => {
-      const dateKey = formatDateValue(date);
-      if (minDateKey && dateKey < minDateKey) {
-        return true;
-      }
-      if (maxDateKey && dateKey > maxDateKey) {
-        return true;
-      }
-      return false;
-    },
-    [maxDateKey, minDateKey],
-  );
-
-  useEffect(() => {
-    if (!open) {
-      return;
-    }
-    setDraftDate(selectedDate);
-    setDraftTime(timeParts.time);
-    setCalendarMonth(resolveInitialCalendarMonth());
-  }, [open, resolveInitialCalendarMonth, selectedDate, timeParts.time]);
-
-  useEffect(() => {
-    if (!open) {
-      return undefined;
-    }
-
-    const handleEscape = (event: KeyboardEvent) => {
-      if (event.key === 'Escape') {
-        setOpen(false);
-      }
-    };
-
-    document.addEventListener('keydown', handleEscape);
-
-    return () => {
-      document.removeEventListener('keydown', handleEscape);
-    };
-  }, [open]);
-
-  const handleApply = () => {
-    if (!draftDate) {
-      return;
-    }
-    onChange(
-      combineDateAndTime(formatDateValue(draftDate), draftTime || defaultTime),
-    );
-    setOpen(false);
-  };
-
-  return (
-    <div className='relative'>
-      <Button
-        size='sm'
-        variant='outline'
-        type='button'
-        aria-label={placeholder}
-        onClick={() => setOpen(current => !current)}
-        className={cn(
-          'h-9 w-full justify-start font-normal',
-          hasValue ? 'pr-16' : 'pr-10',
-        )}
-      >
-        <span
-          className={cn(
-            'flex-1 truncate text-left',
-            value ? 'text-foreground' : 'text-muted-foreground',
-          )}
-        >
-          {label}
-        </span>
-      </Button>
-      {hasValue ? (
-        <button
-          type='button'
-          aria-label={clearLabel}
-          className='absolute right-9 top-1/2 z-10 -translate-y-1/2 rounded-sm p-0.5 text-muted-foreground transition-colors hover:text-foreground'
-          onMouseDown={event => {
-            event.preventDefault();
-            event.stopPropagation();
-          }}
-          onClick={event => {
-            event.preventDefault();
-            event.stopPropagation();
-            onChange('');
-          }}
-        >
-          <X className='h-3.5 w-3.5' />
-        </button>
-      ) : null}
-      <CalendarIcon className='pointer-events-none absolute right-3 top-1/2 h-4 w-4 -translate-y-1/2 text-muted-foreground' />
-      {open ? (
-        <div className='fixed inset-0 z-[70] flex items-center justify-center p-4'>
-          <button
-            type='button'
-            aria-label={clearLabel}
-            className='absolute inset-0 bg-black/20'
-            onClick={() => setOpen(false)}
-          />
-          <div
-            className='relative w-auto max-w-[calc(100vw-2rem)] overflow-auto rounded-md border bg-popover p-0 shadow-md'
-            onClick={event => event.stopPropagation()}
-          >
-            <Calendar
-              mode='single'
-              month={calendarMonth}
-              numberOfMonths={1}
-              selected={draftDate}
-              disabled={isDayDisabled}
-              onMonthChange={setCalendarMonth}
-              onSelect={date => {
-                setDraftDate(date || undefined);
-                if (date) {
-                  setCalendarMonth(date);
-                }
-              }}
-              className='p-3 md:p-4 [--cell-size:2.3rem]'
-            />
-            <div className='border-t border-border px-4 py-3'>
-              <FormField label={timeLabel}>
-                <AdminTimeSelect
-                  value={draftTime}
-                  onChange={setDraftTime}
-                  minTime={minTime}
-                  maxTime={maxTime}
-                  dropdownClassName='bottom-full top-auto mb-1 mt-0'
-                />
-              </FormField>
-            </div>
-            <div className='flex items-center justify-end gap-2 border-t border-border px-3 py-2'>
-              <Button
-                size='sm'
-                variant='ghost'
-                type='button'
-                onClick={() => {
-                  onChange('');
-                  setOpen(false);
-                }}
-              >
-                {resetLabel}
-              </Button>
-              <Button
-                size='sm'
-                type='button'
-                disabled={!draftDate || isDraftTimeOutOfRange}
-                onClick={handleApply}
-              >
-                {t('common.core.confirm')}
-              </Button>
-            </div>
-          </div>
-        </div>
-      ) : null}
-    </div>
-  );
-};
-
-const PromotionCouponCodesDialog = ({
-  open,
-  onOpenChange,
-  couponBid,
-  couponName,
-}: {
-  open: boolean;
-  onOpenChange: (open: boolean) => void;
-  couponBid: string;
-  couponName: string;
-}) => {
-  const { t } = useTranslation();
-  const { t: tPromotion } = useTranslation('module.operationsPromotion');
-  const [loading, setLoading] = useState(false);
-  const [pageIndex, setPageIndex] = useState(1);
-  const [pageCount, setPageCount] = useState(0);
-  const [codes, setCodes] = useState<AdminPromotionCouponCodeItem[]>([]);
-  const [keyword, setKeyword] = useState('');
-  const [appliedKeyword, setAppliedKeyword] = useState('');
-
-  const fetchCodes = useCallback(
-    async (nextPage: number, nextKeyword: string) => {
-      if (!couponBid) {
-        return;
-      }
-      setLoading(true);
-      try {
-        const response = (await api.getAdminOperationPromotionCouponCodes({
-          coupon_bid: couponBid,
-          page_index: nextPage,
-          page_size: PAGE_SIZE,
-          keyword: nextKeyword,
-        })) as AdminPromotionListResponse<AdminPromotionCouponCodeItem>;
-        setCodes(response.items || []);
-        setPageIndex(response.page || nextPage);
-        setPageCount(response.page_count || 0);
-      } catch (error) {
-        setCodes([]);
-        setPageIndex(nextPage);
-        setPageCount(0);
-        showErrorToast(
-          (error as Error).message || tPromotion('messages.loadCodesFailed'),
-        );
-      } finally {
-        setLoading(false);
-      }
-    },
-    [couponBid, tPromotion],
-  );
-
-  useEffect(() => {
-    if (!open || !couponBid) {
-      return;
-    }
-    setKeyword('');
-    setAppliedKeyword('');
-    void fetchCodes(1, '');
-  }, [couponBid, fetchCodes, open]);
-
-  const handleSearch = () => {
-    const nextKeyword = keyword.trim();
-    setAppliedKeyword(nextKeyword);
-    void fetchCodes(1, nextKeyword);
-  };
-
-  return (
-    <Dialog
-      open={open}
-      onOpenChange={onOpenChange}
-    >
-      <DialogContent className='sm:max-w-4xl'>
-        <DialogHeader>
-          <DialogTitle>{tPromotion('coupon.codes')}</DialogTitle>
-        </DialogHeader>
-        <div className='flex max-h-[70vh] min-h-0 flex-col overflow-hidden'>
-          <div className='mb-4 text-sm text-muted-foreground'>
-            {couponName || couponBid}
-          </div>
-          <div className='mb-4 flex items-center gap-3'>
-            <div className='w-full max-w-sm'>
-              <AdminClearableInput
-                value={keyword}
-                onChange={setKeyword}
-                placeholder={tPromotion('coupon.subCodePlaceholder')}
-                clearLabel={t('common.core.close')}
-              />
-            </div>
-            <Button
-              type='button'
-              size='sm'
-              onClick={handleSearch}
-            >
-              {tPromotion('actions.search')}
-            </Button>
-          </div>
-          <AdminTableShell
-            loading={loading}
-            isEmpty={codes.length === 0}
-            emptyContent={tPromotion('messages.emptyCodes')}
-            emptyColSpan={PROMOTION_CODE_DIALOG_COLUMN_COUNT}
-            withTooltipProvider
-            containerClassName='min-h-0 flex-1'
-            tableWrapperClassName='min-h-0 flex-1 overflow-auto'
-            table={emptyRow => (
-              <Table>
-                <TableHeader>
-                  <TableRow>
-                    <TableHead className={TABLE_HEAD_CLASS}>
-                      {tPromotion('coupon.subCode')}
-                    </TableHead>
-                    <TableHead className={TABLE_HEAD_CLASS}>
-                      {tPromotion('table.status')}
-                    </TableHead>
-                    <TableHead className={TABLE_HEAD_CLASS}>
-                      {tPromotion('table.user')}
-                    </TableHead>
-                    <TableHead className={TABLE_HEAD_CLASS}>
-                      {tPromotion('table.orderBid')}
-                    </TableHead>
-                  </TableRow>
-                </TableHeader>
-                <TableBody>
-                  {emptyRow}
-                  {codes.map(item => (
-                    <TableRow key={item.coupon_usage_bid}>
-                      <TableCell className={TABLE_CELL_CLASS}>
-                        {renderTooltipText(item.code)}
-                      </TableCell>
-                      <TableCell className={TABLE_CELL_CLASS}>
-                        {renderTooltipText(
-                          item.status_key ? t(item.status_key) : EMPTY_VALUE,
-                        )}
-                      </TableCell>
-                      <TableCell className={TABLE_CELL_CLASS}>
-                        {renderTooltipText(renderUserLabel(item))}
-                      </TableCell>
-                      <TableCell className={TABLE_LAST_CELL_CLASS}>
-                        {renderTooltipText(item.order_bid)}
-                      </TableCell>
-                    </TableRow>
-                  ))}
-                </TableBody>
-              </Table>
-            )}
-            pagination={{
-              pageIndex,
-              pageCount,
-              onPageChange: page => void fetchCodes(page, appliedKeyword),
-              prevLabel: t('module.order.paginationPrev'),
-              nextLabel: t('module.order.paginationNext'),
-              prevAriaLabel: t('module.order.paginationPrevAriaLabel'),
-              nextAriaLabel: t('module.order.paginationNextAriaLabel'),
-              hideWhenSinglePage: true,
-            }}
-            footerClassName='mt-3'
-          />
-        </div>
-      </DialogContent>
-    </Dialog>
-  );
-};
-
-const PromotionCampaignRedemptionsDialog = ({
-  open,
-  onOpenChange,
-  promoBid,
-  campaignName,
-}: {
-  open: boolean;
-  onOpenChange: (open: boolean) => void;
-  promoBid: string;
-  campaignName: string;
-}) => {
-  const { t } = useTranslation();
-  const { t: tPromotion } = useTranslation('module.operationsPromotion');
-  const [loading, setLoading] = useState(false);
-  const [pageIndex, setPageIndex] = useState(1);
-  const [pageCount, setPageCount] = useState(0);
-  const [redemptions, setRedemptions] = useState<
-    AdminPromotionCampaignRedemptionItem[]
-  >([]);
-
-  const fetchRedemptions = useCallback(
-    async (nextPage: number) => {
-      if (!promoBid) {
-        return;
-      }
-      setLoading(true);
-      try {
-        const response =
-          (await api.getAdminOperationPromotionCampaignRedemptions({
-            promo_bid: promoBid,
-            page_index: nextPage,
-            page_size: PAGE_SIZE,
-          })) as AdminPromotionListResponse<AdminPromotionCampaignRedemptionItem>;
-        setRedemptions(response.items || []);
-        setPageIndex(response.page || nextPage);
-        setPageCount(response.page_count || 0);
-      } catch (error) {
-        setRedemptions([]);
-        setPageIndex(nextPage);
-        setPageCount(0);
-        showErrorToast(
-          (error as Error).message ||
-            tPromotion('messages.loadRedemptionsFailed'),
-        );
-      } finally {
-        setLoading(false);
-      }
-    },
-    [promoBid, tPromotion],
-  );
-
-  useEffect(() => {
-    if (!open || !promoBid) {
-      return;
-    }
-    void fetchRedemptions(1);
-  }, [fetchRedemptions, open, promoBid]);
-
-  return (
-    <Dialog
-      open={open}
-      onOpenChange={onOpenChange}
-    >
-      <DialogContent className='sm:max-w-5xl'>
-        <DialogHeader>
-          <DialogTitle>{tPromotion('campaign.redemptions')}</DialogTitle>
-        </DialogHeader>
-        <div className='flex max-h-[70vh] min-h-0 flex-col overflow-hidden'>
-          <div className='mb-4 text-sm text-muted-foreground'>
-            {campaignName || promoBid}
-          </div>
-          <AdminTableShell
-            loading={loading}
-            isEmpty={redemptions.length === 0}
-            emptyContent={tPromotion('messages.emptyRedemptions')}
-            emptyColSpan={PROMOTION_REDEMPTION_DIALOG_COLUMN_COUNT}
-            withTooltipProvider
-            containerClassName='min-h-0 flex-1'
-            tableWrapperClassName='min-h-0 flex-1 overflow-auto'
-            table={emptyRow => (
-              <Table>
-                <TableHeader>
-                  <TableRow>
-                    <TableHead className={TABLE_HEAD_CLASS}>
-                      {tPromotion('table.appliedAt')}
-                    </TableHead>
-                    <TableHead className={TABLE_HEAD_CLASS}>
-                      {tPromotion('table.user')}
-                    </TableHead>
-                    <TableHead className={TABLE_HEAD_CLASS}>
-                      {tPromotion('table.orderBid')}
-                    </TableHead>
-                    <TableHead className={TABLE_HEAD_CLASS}>
-                      {tPromotion('campaign.discountAmount')}
-                    </TableHead>
-                  </TableRow>
-                </TableHeader>
-                <TableBody>
-                  {emptyRow}
-                  {redemptions.map(item => (
-                    <TableRow key={item.redemption_bid}>
-                      <TableCell className={TABLE_CELL_CLASS}>
-                        {renderTooltipText(
-                          formatAdminUtcDateTime(item.applied_at),
-                        )}
-                      </TableCell>
-                      <TableCell className={TABLE_CELL_CLASS}>
-                        {renderTooltipText(renderUserLabel(item))}
-                      </TableCell>
-                      <TableCell className={TABLE_CELL_CLASS}>
-                        {renderTooltipText(item.order_bid)}
-                      </TableCell>
-                      <TableCell className={TABLE_LAST_CELL_CLASS}>
-                        {renderTooltipText(item.discount_amount)}
-                      </TableCell>
-                    </TableRow>
-                  ))}
-                </TableBody>
-              </Table>
-            )}
-            pagination={{
-              pageIndex,
-              pageCount,
-              onPageChange: page => void fetchRedemptions(page),
-              prevLabel: t('module.order.paginationPrev'),
-              nextLabel: t('module.order.paginationNext'),
-              prevAriaLabel: t('module.order.paginationPrevAriaLabel'),
-              nextAriaLabel: t('module.order.paginationNextAriaLabel'),
-              hideWhenSinglePage: true,
-            }}
-            footerClassName='mt-3'
-          />
-        </div>
-      </DialogContent>
-    </Dialog>
-  );
-};
-
-const PromotionCouponUsageDialog = ({
-  open,
-  onOpenChange,
-  couponBid,
-  couponName,
-  showCourseColumn,
-}: {
-  open: boolean;
-  onOpenChange: (open: boolean) => void;
-  couponBid: string;
-  couponName: string;
-  showCourseColumn: boolean;
-}) => {
-  const { t } = useTranslation();
-  const { t: tPromotion } = useTranslation('module.operationsPromotion');
-  const [loading, setLoading] = useState(false);
-  const [pageIndex, setPageIndex] = useState(1);
-  const [pageCount, setPageCount] = useState(0);
-  const [usages, setUsages] = useState<AdminPromotionCouponUsageItem[]>([]);
-
-  const fetchUsages = useCallback(
-    async (nextPage: number) => {
-      if (!couponBid) {
-        return;
-      }
-      setLoading(true);
-      try {
-        const response = (await api.getAdminOperationPromotionCouponUsages({
-          coupon_bid: couponBid,
-          page_index: nextPage,
-          page_size: PAGE_SIZE,
-        })) as AdminPromotionListResponse<AdminPromotionCouponUsageItem>;
-        setUsages(response.items || []);
-        setPageIndex(response.page || nextPage);
-        setPageCount(response.page_count || 0);
-      } catch (error) {
-        setUsages([]);
-        setPageIndex(nextPage);
-        setPageCount(0);
-        showErrorToast(
-          (error as Error).message || tPromotion('messages.loadUsagesFailed'),
-        );
-      } finally {
-        setLoading(false);
-      }
-    },
-    [couponBid, tPromotion],
-  );
-
-  useEffect(() => {
-    if (!open || !couponBid) {
-      return;
-    }
-    void fetchUsages(1);
-  }, [couponBid, fetchUsages, open]);
-
-  return (
-    <Dialog
-      open={open}
-      onOpenChange={onOpenChange}
-    >
-      <DialogContent className='sm:max-w-4xl'>
-        <DialogHeader>
-          <DialogTitle>{tPromotion('coupon.usages')}</DialogTitle>
-        </DialogHeader>
-        <div className='flex max-h-[70vh] min-h-0 flex-col overflow-hidden'>
-          <div className='mb-4 text-sm text-muted-foreground'>
-            {couponName || couponBid}
-          </div>
-          <AdminTableShell
-            loading={loading}
-            isEmpty={usages.length === 0}
-            emptyContent={tPromotion('messages.emptyUsages')}
-            emptyColSpan={
-              showCourseColumn
-                ? PROMOTION_USAGE_DIALOG_COLUMN_COUNT.withCourse
-                : PROMOTION_USAGE_DIALOG_COLUMN_COUNT.default
-            }
-            withTooltipProvider
-            containerClassName='min-h-0 flex-1'
-            tableWrapperClassName='min-h-0 flex-1 overflow-auto'
-            table={emptyRow => (
-              <Table containerClassName='overflow-visible max-h-none'>
-                <TableHeader>
-                  <TableRow>
-                    <TableHead className={TABLE_HEAD_CLASS}>
-                      {tPromotion('table.usedAt')}
-                    </TableHead>
-                    <TableHead className={TABLE_HEAD_CLASS}>
-                      {tPromotion('coupon.code')}
-                    </TableHead>
-                    {showCourseColumn ? (
-                      <TableHead className={TABLE_HEAD_CLASS}>
-                        {tPromotion('table.redeemedCourse')}
-                      </TableHead>
-                    ) : null}
-                    <TableHead className={TABLE_HEAD_CLASS}>
-                      {tPromotion('table.user')}
-                    </TableHead>
-                    <TableHead className={TABLE_HEAD_CLASS}>
-                      {tPromotion('table.orderBid')}
-                    </TableHead>
-                  </TableRow>
-                </TableHeader>
-                <TableBody>
-                  {emptyRow}
-                  {usages.map(item => (
-                    <TableRow key={item.coupon_usage_bid}>
-                      <TableCell className={TABLE_CELL_CLASS}>
-                        {renderTooltipText(
-                          formatAdminUtcDateTime(item.used_at),
-                        )}
-                      </TableCell>
-                      <TableCell className={TABLE_CELL_CLASS}>
-                        {renderTooltipText(item.code)}
-                      </TableCell>
-                      {showCourseColumn ? (
-                        <TableCell className={TABLE_CELL_CLASS}>
-                          {renderTooltipText(
-                            item.course_name || item.shifu_bid || EMPTY_VALUE,
-                          )}
-                        </TableCell>
-                      ) : null}
-                      <TableCell className={TABLE_CELL_CLASS}>
-                        {renderTooltipText(renderUserLabel(item))}
-                      </TableCell>
-                      <TableCell className={TABLE_LAST_CELL_CLASS}>
-                        {renderTooltipText(item.order_bid)}
-                      </TableCell>
-                    </TableRow>
-                  ))}
-                </TableBody>
-              </Table>
-            )}
-            pagination={{
-              pageIndex,
-              pageCount,
-              onPageChange: page => void fetchUsages(page),
-              prevLabel: t('module.order.paginationPrev'),
-              nextLabel: t('module.order.paginationNext'),
-              prevAriaLabel: t('module.order.paginationPrevAriaLabel'),
-              nextAriaLabel: t('module.order.paginationNextAriaLabel'),
-              hideWhenSinglePage: true,
-            }}
-            footerClassName='mt-3'
-          />
-        </div>
-      </DialogContent>
-    </Dialog>
-  );
-};
-
-const PromotionCouponDialog = ({
-  open,
-  onOpenChange,
-  onSubmit,
-  coupon,
-}: {
-  open: boolean;
-  onOpenChange: (open: boolean) => void;
-  onSubmit: (payload: CouponFormState) => Promise<void>;
-  coupon?: AdminPromotionCouponItem | null;
-}) => {
-  const { t } = useTranslation();
-  const { t: tPromotion } = useTranslation('module.operationsPromotion');
-  const [form, setForm] = useState<CouponFormState>(() =>
-    createDefaultCouponForm(),
-  );
-  const [submitting, setSubmitting] = useState(false);
-
-  useEffect(() => {
-    if (open) {
-      setForm(
-        coupon ? createCouponFormFromItem(coupon) : createDefaultCouponForm(),
-      );
-    }
-  }, [coupon, open]);
-
-  const isEditing = Boolean(coupon);
-  const isSingleUseCoupon = form.usage_type === '802';
-  const isPercentDiscount = form.discount_type === '702';
-  const valueLabel = isPercentDiscount
-    ? tPromotion('coupon.valuePercent')
-    : tPromotion('coupon.valueAmount');
-  const valuePlaceholder = isPercentDiscount
-    ? tPromotion('coupon.valuePercentPlaceholder')
-    : tPromotion('coupon.valueAmountPlaceholder');
-
-  const handleSubmit = async () => {
-    const normalizedName = form.name.trim();
-    const normalizedCode = form.code.trim();
-    const normalizedQuantity = form.total_count.trim();
-    const normalizedCourseId = form.shifu_bid.trim();
-    const normalizedValue = form.value.trim();
-    const startAtDate = parseLocalDateTimeInput(form.start_at);
-    const endAtDate = parseLocalDateTimeInput(form.end_at);
-
-    if (!normalizedName) {
-      showDefaultToast(tPromotion('validation.couponNameRequired'));
-      return;
-    }
-    if (!form.usage_type) {
-      showDefaultToast(tPromotion('validation.usageTypeRequired'));
-      return;
-    }
-    if (!form.discount_type) {
-      showDefaultToast(tPromotion('validation.discountTypeRequired'));
-      return;
-    }
-    if (!normalizedValue) {
-      showDefaultToast(
-        isPercentDiscount
-          ? tPromotion('validation.valuePercentRequired')
-          : tPromotion('validation.valueAmountRequired'),
-      );
-      return;
-    }
-
-    const numericValue = Number(normalizedValue);
-    if (!Number.isFinite(numericValue)) {
-      showDefaultToast(
-        isPercentDiscount
-          ? tPromotion('validation.valuePercentInvalid')
-          : tPromotion('validation.valueAmountInvalid'),
-      );
-      return;
-    }
-    if (isPercentDiscount) {
-      if (numericValue <= 0 || numericValue > 100) {
-        showDefaultToast(tPromotion('validation.valuePercentInvalid'));
-        return;
-      }
-    } else if (numericValue <= 0) {
-      showDefaultToast(tPromotion('validation.valueAmountInvalid'));
-      return;
-    }
-
-    if (!isSingleUseCoupon && !normalizedCode) {
-      showDefaultToast(tPromotion('validation.codeRequired'));
-      return;
-    }
-    if (!normalizedQuantity) {
-      showDefaultToast(tPromotion('validation.quantityRequired'));
-      return;
-    }
-    if (
-      !isPositiveIntegerString(normalizedQuantity) ||
-      Number(normalizedQuantity) <= 0
-    ) {
-      showDefaultToast(tPromotion('validation.quantityInvalid'));
-      return;
-    }
-    if (form.scope_type === 'single_course' && !normalizedCourseId) {
-      showDefaultToast(tPromotion('validation.courseIdRequired'));
-      return;
-    }
-    if (!form.start_at) {
-      showDefaultToast(tPromotion('validation.startAtRequired'));
-      return;
-    }
-    if (!form.end_at) {
-      showDefaultToast(tPromotion('validation.endAtRequired'));
-      return;
-    }
-    if (
-      !startAtDate ||
-      !endAtDate ||
-      endAtDate.getTime() < startAtDate.getTime()
-    ) {
-      showDefaultToast(tPromotion('validation.endAtInvalid'));
-      return;
-    }
-
-    setSubmitting(true);
-    try {
-      await onSubmit(form);
-      onOpenChange(false);
-    } catch (error) {
-      showErrorToast((error as Error).message || t('common.core.submitFailed'));
-    } finally {
-      setSubmitting(false);
-    }
-  };
-
-  return (
-    <Dialog
-      open={open}
-      onOpenChange={onOpenChange}
-    >
-      <DialogContent className='sm:max-w-[680px]'>
-        <DialogHeader>
-          <DialogTitle>
-            {isEditing
-              ? tPromotion('coupon.editDialogTitle')
-              : tPromotion('coupon.dialogTitle')}
-          </DialogTitle>
-        </DialogHeader>
-        <div className='grid gap-4 md:grid-cols-2'>
-          <FormField label={tPromotion('table.name')}>
-            <Input
-              className='h-9'
-              value={form.name}
-              placeholder={tPromotion('filters.namePlaceholder')}
-              onChange={event =>
-                setForm(current => ({ ...current, name: event.target.value }))
-              }
-            />
-          </FormField>
-          <FormField label={tPromotion('table.usageType')}>
-            <Select
-              value={form.usage_type}
-              onValueChange={value =>
-                setForm(current => ({
-                  ...current,
-                  usage_type: value,
-                  code: value === '801' ? current.code : '',
-                }))
-              }
-              disabled={isEditing}
-            >
-              <SelectTrigger className='h-9'>
-                <SelectValue placeholder={tPromotion('filters.usageType')} />
-              </SelectTrigger>
-              <SelectContent>
-                <SelectItem value='801'>
-                  {tPromotion('usageType.generic')}
-                </SelectItem>
-                <SelectItem value='802'>
-                  {tPromotion('usageType.singleUse')}
-                </SelectItem>
-              </SelectContent>
-            </Select>
-          </FormField>
-          <FormField label={tPromotion('filters.discountType')}>
-            <Select
-              value={form.discount_type}
-              onValueChange={value =>
-                setForm(current => ({ ...current, discount_type: value }))
-              }
-              disabled={isEditing}
-            >
-              <SelectTrigger className='h-9'>
-                <SelectValue placeholder={tPromotion('filters.discountType')} />
-              </SelectTrigger>
-              <SelectContent>
-                <SelectItem value='701'>
-                  {tPromotion('discountType.fixed')}
-                </SelectItem>
-                <SelectItem value='702'>
-                  {tPromotion('discountType.percent')}
-                </SelectItem>
-              </SelectContent>
-            </Select>
-          </FormField>
-          <FormField label={valueLabel}>
-            <Input
-              className='h-9'
-              value={form.value}
-              placeholder={valuePlaceholder}
-              onChange={event =>
-                setForm(current => ({ ...current, value: event.target.value }))
-              }
-              disabled={isEditing}
-            />
-          </FormField>
-          {isSingleUseCoupon ? null : (
-            <FormField label={tPromotion('coupon.code')}>
-              <Input
-                className='h-9'
-                value={form.code}
-                placeholder={tPromotion('coupon.codePlaceholder')}
-                onChange={event =>
-                  setForm(current => ({ ...current, code: event.target.value }))
-                }
-                disabled={isEditing}
-              />
-            </FormField>
-          )}
-          <FormField label={tPromotion('coupon.quantity')}>
-            <Input
-              className='h-9'
-              value={form.total_count}
-              placeholder={tPromotion('coupon.quantityPlaceholder')}
-              onChange={event =>
-                setForm(current => ({
-                  ...current,
-                  total_count: event.target.value,
-                }))
-              }
-            />
-          </FormField>
-          <FormField label={tPromotion('coupon.scopeType')}>
-            <Select
-              value={form.scope_type}
-              onValueChange={value =>
-                setForm(current => ({
-                  ...current,
-                  scope_type: value,
-                  shifu_bid: value === 'single_course' ? current.shifu_bid : '',
-                }))
-              }
-              disabled={isEditing}
-            >
-              <SelectTrigger className='h-9'>
-                <SelectValue />
-              </SelectTrigger>
-              <SelectContent>
-                <SelectItem value='all_courses'>
-                  {tPromotion('scope.allCourses')}
-                </SelectItem>
-                <SelectItem value='single_course'>
-                  {tPromotion('scope.singleCourse')}
-                </SelectItem>
-              </SelectContent>
-            </Select>
-          </FormField>
-          <FormField label={tPromotion('coupon.courseId')}>
-            <Input
-              className='h-9'
-              value={form.shifu_bid}
-              placeholder={tPromotion('filters.courseIdPlaceholder')}
-              onChange={event =>
-                setForm(current => ({
-                  ...current,
-                  shifu_bid: event.target.value,
-                }))
-              }
-              disabled={isEditing || form.scope_type !== 'single_course'}
-            />
-          </FormField>
-          <FormField label={tPromotion('coupon.startAt')}>
-            <PromotionDateTimePicker
-              value={form.start_at}
-              placeholder={tPromotion('coupon.startAt')}
-              resetLabel={t('module.order.filters.reset')}
-              clearLabel={t('common.core.close')}
-              timeLabel={tPromotion('coupon.startAt')}
-              defaultTime={DEFAULT_START_TIME}
-              maxDateTime={form.end_at}
-              onChange={nextValue =>
-                setForm(current => ({
-                  ...current,
-                  start_at: nextValue,
-                }))
-              }
-            />
-          </FormField>
-          <FormField label={tPromotion('coupon.endAt')}>
-            <PromotionDateTimePicker
-              value={form.end_at}
-              placeholder={tPromotion('coupon.endAt')}
-              resetLabel={t('module.order.filters.reset')}
-              clearLabel={t('common.core.close')}
-              timeLabel={tPromotion('coupon.endAt')}
-              defaultTime={DEFAULT_END_TIME}
-              minDateTime={form.start_at}
-              onChange={nextValue =>
-                setForm(current => ({
-                  ...current,
-                  end_at: nextValue,
-                }))
-              }
-            />
-          </FormField>
-        </div>
-        <DialogFooter>
-          <Button
-            type='button'
-            variant='outline'
-            onClick={() => onOpenChange(false)}
-          >
-            {t('common.core.cancel')}
-          </Button>
-          <Button
-            type='button'
-            onClick={() => void handleSubmit()}
-            disabled={submitting}
-          >
-            {isEditing
-              ? tPromotion('actions.confirmUpdate')
-              : tPromotion('actions.confirmCreate')}
-          </Button>
-        </DialogFooter>
-      </DialogContent>
-    </Dialog>
-  );
-};
-
-const PromotionCampaignDialog = ({
-  open,
-  onOpenChange,
-  onSubmit,
-  campaign,
-  strategyEditable,
-}: {
-  open: boolean;
-  onOpenChange: (open: boolean) => void;
-  onSubmit: (payload: CampaignFormState) => Promise<void>;
-  campaign?: {
-    item: AdminPromotionCampaignItem;
-    description: string;
-  } | null;
-  strategyEditable?: boolean;
-}) => {
-  const { t } = useTranslation();
-  const { t: tPromotion } = useTranslation('module.operationsPromotion');
-  const [form, setForm] = useState<CampaignFormState>(() =>
-    createDefaultCampaignForm(),
-  );
-  const [submitting, setSubmitting] = useState(false);
-
-  useEffect(() => {
-    if (open) {
-      setForm(
-        campaign
-          ? createCampaignFormFromItem(campaign.item, campaign.description)
-          : createDefaultCampaignForm(),
-      );
-    }
-  }, [campaign, open]);
-
-  const isEditing = Boolean(campaign);
-
-  const isPercentDiscount = form.discount_type === '702';
-  const valueLabel = form.discount_type
-    ? isPercentDiscount
-      ? tPromotion('coupon.valuePercent')
-      : tPromotion('coupon.valueAmount')
-    : tPromotion('campaign.value');
-  const valuePlaceholder = form.discount_type
-    ? isPercentDiscount
-      ? tPromotion('coupon.valuePercentPlaceholder')
-      : tPromotion('coupon.valueAmountPlaceholder')
-    : tPromotion('campaign.valuePlaceholder');
-
-  const handleSubmit = async () => {
-    const normalizedName = form.name.trim();
-    const normalizedCourseId = form.shifu_bid.trim();
-    const normalizedValue = form.value.trim();
-    const startAtDate = parseLocalDateTimeInput(form.start_at);
-    const endAtDate = parseLocalDateTimeInput(form.end_at);
-
-    if (!normalizedName) {
-      showDefaultToast(tPromotion('validation.campaignNameRequired'));
-      return;
-    }
-    if (!form.apply_type) {
-      showDefaultToast(tPromotion('validation.campaignApplyTypeRequired'));
-      return;
-    }
-    if (!normalizedCourseId) {
-      showDefaultToast(tPromotion('validation.courseIdRequired'));
-      return;
-    }
-    if (!form.discount_type) {
-      showDefaultToast(tPromotion('validation.discountTypeRequired'));
-      return;
-    }
-    if (!normalizedValue) {
-      showDefaultToast(
-        isPercentDiscount
-          ? tPromotion('validation.valuePercentRequired')
-          : tPromotion('validation.valueAmountRequired'),
-      );
-      return;
-    }
-    const numericValue = Number(normalizedValue);
-    if (!Number.isFinite(numericValue)) {
-      showDefaultToast(
-        isPercentDiscount
-          ? tPromotion('validation.valuePercentInvalid')
-          : tPromotion('validation.valueAmountInvalid'),
-      );
-      return;
-    }
-    if (isPercentDiscount) {
-      if (numericValue <= 0 || numericValue > 100) {
-        showDefaultToast(tPromotion('validation.valuePercentInvalid'));
-        return;
-      }
-    } else if (numericValue <= 0) {
-      showDefaultToast(tPromotion('validation.valueAmountInvalid'));
-      return;
-    }
-    if (!form.start_at) {
-      showDefaultToast(tPromotion('validation.startAtRequired'));
-      return;
-    }
-    if (!form.end_at) {
-      showDefaultToast(tPromotion('validation.endAtRequired'));
-      return;
-    }
-    if (
-      !startAtDate ||
-      !endAtDate ||
-      endAtDate.getTime() < startAtDate.getTime()
-    ) {
-      showDefaultToast(tPromotion('validation.endAtInvalid'));
-      return;
-    }
-
-    setSubmitting(true);
-    try {
-      await onSubmit(form);
-      onOpenChange(false);
-    } catch (error) {
-      showErrorToast((error as Error).message || t('common.core.submitFailed'));
-    } finally {
-      setSubmitting(false);
-    }
-  };
-
-  return (
-    <Dialog
-      open={open}
-      onOpenChange={onOpenChange}
-    >
-      <DialogContent className='sm:max-w-[700px]'>
-        <DialogHeader>
-          <DialogTitle>
-            {isEditing
-              ? tPromotion('campaign.editDialogTitle')
-              : tPromotion('campaign.dialogTitle')}
-          </DialogTitle>
-        </DialogHeader>
-        <div className='grid gap-4 md:grid-cols-2'>
-          <FormField label={tPromotion('table.campaignName')}>
-            <Input
-              className='h-9'
-              value={form.name}
-              placeholder={tPromotion('campaign.namePlaceholder')}
-              onChange={event =>
-                setForm(current => ({ ...current, name: event.target.value }))
-              }
-            />
-          </FormField>
-          <FormField label={tPromotion('coupon.courseId')}>
-            <Input
-              className='h-9'
-              value={form.shifu_bid}
-              placeholder={tPromotion('filters.courseIdPlaceholder')}
-              onChange={event =>
-                setForm(current => ({
-                  ...current,
-                  shifu_bid: event.target.value,
-                }))
-              }
-              disabled={isEditing}
-            />
-          </FormField>
-          <FormField label={tPromotion('campaign.applyType')}>
-            <Select
-              value={form.apply_type}
-              onValueChange={value =>
-                setForm(current => ({ ...current, apply_type: value }))
-              }
-              disabled={isEditing && !strategyEditable}
-            >
-              <SelectTrigger className='h-9'>
-                <SelectValue
-                  placeholder={tPromotion('campaign.applyTypePlaceholder')}
-                />
-              </SelectTrigger>
-              <SelectContent>
-                <SelectItem value='2101'>
-                  {tPromotion('campaign.applyTypeAuto')}
-                </SelectItem>
-                <SelectItem value='2102'>
-                  {tPromotion('campaign.applyTypeEvent')}
-                </SelectItem>
-                <SelectItem value='2103'>
-                  {tPromotion('campaign.applyTypeManual')}
-                </SelectItem>
-              </SelectContent>
-            </Select>
-          </FormField>
-          <FormField label={tPromotion('campaign.channel')}>
-            <Input
-              className='h-9'
-              value={form.channel}
-              placeholder={tPromotion('campaign.channelPlaceholder')}
-              onChange={event =>
-                setForm(current => ({
-                  ...current,
-                  channel: event.target.value,
-                }))
-              }
-              disabled={isEditing}
-            />
-          </FormField>
-          <FormField label={tPromotion('filters.discountType')}>
-            <Select
-              value={form.discount_type}
-              onValueChange={value =>
-                setForm(current => ({ ...current, discount_type: value }))
-              }
-              disabled={isEditing}
-            >
-              <SelectTrigger className='h-9'>
-                <SelectValue placeholder={tPromotion('filters.discountType')} />
-              </SelectTrigger>
-              <SelectContent>
-                <SelectItem value='701'>
-                  {tPromotion('discountType.fixed')}
-                </SelectItem>
-                <SelectItem value='702'>
-                  {tPromotion('discountType.percent')}
-                </SelectItem>
-              </SelectContent>
-            </Select>
-          </FormField>
-          <FormField label={valueLabel}>
-            <Input
-              className='h-9'
-              value={form.value}
-              placeholder={valuePlaceholder}
-              onChange={event =>
-                setForm(current => ({ ...current, value: event.target.value }))
-              }
-              disabled={isEditing}
-            />
-          </FormField>
-          <FormField label={tPromotion('campaign.startAt')}>
-            <PromotionDateTimePicker
-              value={form.start_at}
-              placeholder={tPromotion('campaign.startAtPlaceholder')}
-              resetLabel={t('module.order.filters.reset')}
-              clearLabel={t('common.core.close')}
-              timeLabel={tPromotion('campaign.startAt')}
-              defaultTime={DEFAULT_START_TIME}
-              maxDateTime={form.end_at}
-              onChange={nextValue =>
-                setForm(current => ({
-                  ...current,
-                  start_at: nextValue,
-                }))
-              }
-            />
-          </FormField>
-          <FormField label={tPromotion('campaign.endAt')}>
-            <PromotionDateTimePicker
-              value={form.end_at}
-              placeholder={tPromotion('campaign.endAtPlaceholder')}
-              resetLabel={t('module.order.filters.reset')}
-              clearLabel={t('common.core.close')}
-              timeLabel={tPromotion('campaign.endAt')}
-              defaultTime={DEFAULT_END_TIME}
-              minDateTime={form.start_at}
-              onChange={nextValue =>
-                setForm(current => ({
-                  ...current,
-                  end_at: nextValue,
-                }))
-              }
-            />
-          </FormField>
-          <div className='md:col-span-2'>
-            <FormField label={tPromotion('campaign.description')}>
-              <Textarea
-                value={form.description}
-                placeholder={tPromotion('campaign.descriptionPlaceholder')}
-                onChange={event =>
-                  setForm(current => ({
-                    ...current,
-                    description: event.target.value,
-                  }))
-                }
-              />
-            </FormField>
-          </div>
-        </div>
-        <DialogFooter>
-          <Button
-            type='button'
-            variant='outline'
-            onClick={() => onOpenChange(false)}
-          >
-            {t('common.core.cancel')}
-          </Button>
-          <Button
-            type='button'
-            onClick={() => void handleSubmit()}
-            disabled={submitting}
-          >
-            {isEditing
-              ? tPromotion('actions.confirmUpdate')
-              : tPromotion('actions.confirmCreate')}
-          </Button>
-        </DialogFooter>
-      </DialogContent>
-    </Dialog>
-  );
-};
+import {
+  PromotionCampaignDialog,
+  PromotionCouponDialog,
+} from './PromotionFormDialogs';
+import {
+  PromotionCampaignRedemptionsDialog,
+  PromotionCouponCodesDialog,
+  PromotionCouponUsageDialog,
+} from './PromotionRecordDialogs';
+import PromotionStatusConfirmDialog from './PromotionStatusConfirmDialog';
+import {
+  ALL_OPTION_VALUE,
+  CAMPAIGN_COLUMN_WIDTH_STORAGE_KEY,
+  CAMPAIGN_DEFAULT_COLUMN_WIDTHS,
+  type CampaignColumnKey,
+  type CampaignFilters,
+  type CampaignFormState,
+  COLUMN_MAX_WIDTH,
+  COLUMN_MIN_WIDTH,
+  COUPON_COLUMN_WIDTH_STORAGE_KEY,
+  COUPON_DEFAULT_COLUMN_WIDTHS,
+  type CouponColumnKey,
+  type CouponFilters,
+  type CouponFormState,
+  createDefaultCampaignFilters,
+  createDefaultCouponFilters,
+  EMPTY_VALUE,
+  type ErrorState,
+  PAGE_SIZE,
+  type PromotionStatusChangeTarget,
+  type PromotionTab,
+  renderCouponAttentionBadges,
+  renderPromotionStatusBadge,
+  renderRuleLabel,
+  renderTimeRange,
+  renderTooltipText,
+  resolveCampaignApplyTypeLabel,
+  resolveCouponScopeLabel,
+  resolveCouponUsageTypeLabel,
+  SectionCard,
+  shouldShowCampaignStatusToggle,
+  shouldShowCouponStatusToggle,
+  SINGLE_SELECT_ITEM_CLASS,
+  TABLE_ACTION_CELL_CLASS,
+  TABLE_ACTION_HEAD_CLASS,
+  TABLE_CELL_CLASS,
+  TABLE_HEAD_CLASS,
+  TABLE_LAST_CELL_CLASS,
+  downloadExcelCompatibleCodesFile,
+  canEditCampaignStrategyFields,
+} from './promotionPageShared';
 
 export default function AdminOperationPromotionsPage() {
   const { t } = useTranslation();
